Bug 1739443: Prune mercurial.ini r=glandium,sheehan
Most of our "opinionated setup" is now happening via
`hg configwizard` in `version-control-tools`. This affects
`extensions`, some `diff` config, and other pieces.

A bunch of `[merge-tools]` lines were equivalent with the default
config in `mercurial.defaultrc.mergetools.rc`, so are redundant
and are removed.

The only remaining config that should still be shipped
with MozillaBuild's `mercurial.ini` should be
MozillaBuild-specific (or carried along for historical/compatibility
reasons, such as the remaining kdiff3 options that are slightly
different than the mercurial defaults).

Differential Revision: https://phabricator.services.mozilla.com/D134848