Bug 1751500: Work around erroneous line breaks in copied text r=glandium
ConPTY is still new enough to have some rough edges, and one of
them is that it adds line breaks to copied text that is wrapped
in the console.

To work around this, we fall back to using the Command Prompt as
a terminal again (instead of MinTTY, just as we did for previous
MozillaBuild releases).

This required two additional tweaks:
* The setting of `TERM` to avoid HG colouring issues
* The usage of the default `.inputrc` to avoid bell noises
  when autocompleting directories, etc.

Differential Revision: https://phabricator.services.mozilla.com/D136674