Bug 1751500: Re-add support for the `USE_MINTTY` option r=glandium
The `USE_MINTTY` option was originally removed as part of Bug 1725895,
since Mintty was going to be used by default.

However, an earlier patch (part of bug 1751500) had to revert the
terminal change and make the Command Prompt the default terminal again.

Since the terminal situation is back to how it was with MozillaBuild
3.4, we might as well continue to export `USE_MINTTY` to avoid a
breaking change. For what it's worth, Mintty is //more// viable today
than it was with MozillaBuild 3.4 due to ConPTY and the updated MSYS2
environment, so keeping it accessible is a good idea.

Related, as part of this: move the conpty-enabling line to only happen
if Mintty is used, because otherwise it's unnecessary.

Differential Revision: https://phabricator.services.mozilla.com/D137799