Bug 1753751: Fix vim-quitting not fully cleaning up terminal r=glandium
The vim-clearing issue was caused by our changing of `$TERM` to be
`cygwin` instead of the upstream default of `xterm-256color`.

That `$TERM` change was to avoid an `hg` bug [1] around detecting the
color mode, which assumed that any terminal that had `"xterm"` in the
`$TERM` was coloured with ansi sequences.

Work around this by setting `$TERM` to the old MSYS value
of "cygwin".

Note that this doesn't break compatibility with Mintty: `hg` output
appears to still be properly coloured.

[1] https://bz.mercurial-scm.org/show_bug.cgi?id=6640

Differential Revision: https://phabricator.services.mozilla.com/D137906