Bug 1760862: Fix hg completions not working due to alias r=ahochheiden
HG's bash completions don't work if `hg` is aliased, and that alias has
a space.

Additionally, the only way to influence `git`'s editor (while still
respecting `ui.editor`) is with the `EDITOR` environment variable.

Fortunately, `git`'s bash completions //do// behave nicely with a space,
so move the the `EDITOR` variable declaration to happen in a `git` alias
instead.

Depends on D141801

Differential Revision: https://phabricator.services.mozilla.com/D141802